Merge current state of UDB

See merge request STJr/UltimateZoneBuilder!2

+++ Source/Plugins/UDBScript/BuilderPlug.cs
@@ -30,6 +30,7 @@ using System.Dynamic;
 using System.Windows.Forms;
 using System.IO;
 using System.Linq;
+using System.Numerics;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Text.RegularExpressions;
@@ -49,14 +50,16 @@ namespace CodeImp.DoomBuilder.UDBScript
 {
 	internal class ScriptDirectoryStructure
 	{
+		public string Path;
 		public string Name;
 		public bool Expanded;
 		public string Hash;
 		public List<ScriptDirectoryStructure> Directories;
 		public List<ScriptInfo> Scripts;
 
-		public ScriptDirectoryStructure(string name, bool expanded, string hash)
+		public ScriptDirectoryStructure(string path, string name, bool expanded, string hash)
 		{
+			Path = path;
 			Name = name;
 			Expanded = expanded;
 			Hash = hash;
@@ -70,7 +73,7 @@ namespace CodeImp.DoomBuilder.UDBScript
 		#region ================== Constants
 
 		private static readonly string SCRIPT_FOLDER = "udbscript";
-		public static readonly uint UDB_SCRIPT_VERSION = 4;
+		public static readonly uint UDB_SCRIPT_VERSION = 5;
 
 		#endregion
 
@@ -130,13 +133,18 @@ namespace CodeImp.DoomBuilder.UDBScript
 
 			General.Actions.BindMethods(this);
 
-			watcher = new FileSystemWatcher(Path.Combine(General.AppPath, SCRIPT_FOLDER, "scripts"));
-			watcher.NotifyFilter = NotifyFilters.CreationTime | NotifyFilters.DirectoryName | NotifyFilters.FileName | NotifyFilters.LastWrite | NotifyFilters.Size;
-			watcher.IncludeSubdirectories = true;
-			watcher.Changed += OnWatcherEvent;
-			watcher.Created += OnWatcherEvent;
-			watcher.Deleted += OnWatcherEvent;
-			watcher.Renamed += OnWatcherEvent;
+			string scriptspath = Path.Combine(General.AppPath, SCRIPT_FOLDER, "scripts");
+
+			if (Directory.Exists(scriptspath))
+			{
+				watcher = new FileSystemWatcher(Path.Combine(General.AppPath, SCRIPT_FOLDER, "scripts"));
+				watcher.NotifyFilter = NotifyFilters.CreationTime | NotifyFilters.DirectoryName | NotifyFilters.FileName | NotifyFilters.LastWrite | NotifyFilters.Size;
+				watcher.IncludeSubdirectories = true;
+				watcher.Changed += OnWatcherEvent;
+				watcher.Created += OnWatcherEvent;
+				watcher.Deleted += OnWatcherEvent;
+				watcher.Renamed += OnWatcherEvent;
+			}
 
 			editorexepath = General.Settings.ReadPluginSetting("externaleditor", string.Empty);
 
@@ -152,7 +160,8 @@ namespace CodeImp.DoomBuilder.UDBScript
 			// Methods called by LoadScripts might sleep for some time, so call LoadScripts asynchronously
 			new Task(LoadScripts).Start();
 
-			watcher.EnableRaisingEvents = true;
+			if (watcher != null)
+				watcher.EnableRaisingEvents = true;
 		}
 
 		public override void OnMapOpenEnd()
@@ -162,12 +171,14 @@ namespace CodeImp.DoomBuilder.UDBScript
 			// Methods called by LoadScripts might sleep for some time, so call LoadScripts asynchronously
 			new Task(LoadScripts).Start();
 
-			watcher.EnableRaisingEvents = true;
+			if (watcher != null)
+				watcher.EnableRaisingEvents = true;
 		}
 
 		public override void OnMapCloseBegin()
 		{
-			watcher.EnableRaisingEvents = false;
+			if (watcher != null)
+				watcher.EnableRaisingEvents = false;
 
 			SaveScriptSlotsAndOptions();
 			SaveScriptDirectoryExpansionStatus(scriptdirectorystructure);
@@ -228,6 +239,9 @@ namespace CodeImp.DoomBuilder.UDBScript
 
 		internal void SaveScriptDirectoryExpansionStatus(ScriptDirectoryStructure root)
 		{
+			if (root == null)
+				return;
+
 			if(root.Expanded)
 			{
 				General.Settings.DeletePluginSetting("directoryexpand." + root.Hash);
@@ -386,7 +400,7 @@ namespace CodeImp.DoomBuilder.UDBScript
 			string hash = SHA256Hash.Get(path);
 			bool expanded = General.Settings.ReadPluginSetting("directoryexpand." + hash, true);
 			string name = path.TrimEnd(Path.DirectorySeparatorChar).Split(Path.DirectorySeparatorChar).Last();
-			ScriptDirectoryStructure sds = new ScriptDirectoryStructure(name, expanded, hash);
+			ScriptDirectoryStructure sds = new ScriptDirectoryStructure(path, name, expanded, hash);
 
 			foreach (string directory in Directory.GetDirectories(path))
 				sds.Directories.Add(LoadScriptDirectoryStructure(directory));
@@ -452,41 +466,44 @@ namespace CodeImp.DoomBuilder.UDBScript
 			panel.EndEdit();
 		}
 
-		internal object GetVectorFromObject(object data, bool allow3d)
+		internal Vector3D GetVector3DFromObject(object data)
 		{
 			if (data is Vector2D)
 				return (Vector2D)data;
 			else if (data is Vector2DWrapper)
 				return new Vector2D(((Vector2DWrapper)data)._x, ((Vector2DWrapper)data)._y);
+			else if (data is Vector3D)
+				return (Vector3D)data;
 			else if (data is Vector3DWrapper)
-			{
-				if(allow3d)
-					return new Vector3D(((Vector3DWrapper)data)._x, ((Vector3DWrapper)data)._y, ((Vector3DWrapper)data)._z);
-				else
-					return new Vector2D(((Vector3DWrapper)data)._x, ((Vector3DWrapper)data)._y);
-			}
+				return new Vector3D(((Vector3DWrapper)data)._x, ((Vector3DWrapper)data)._y, ((Vector3DWrapper)data)._z);
 			else if (data.GetType().IsArray)
-			//else if(data is double[])
 			{
-				object[] vals = (object[])data;
-				//double[] vals = (double[])data;
+				object[] rawvals = (object[])data;
+				List<double> vals = new List<double>(rawvals.Length);
 
-				// Make sure all values in the array are doubles
-				foreach (object v in vals)
-					if (!(v is double))
+				// Make sure all values in the array are doubles or BigIntegers
+				foreach (object rv in rawvals)
+				{
+					if (!(rv is double || rv is BigInteger))
 						throw new CantConvertToVectorException("Values in array must be numbers.");
 
-				if (vals.Length == 2)
-					return new Vector2D((double)vals[0], (double)vals[1]);
-				if (vals.Length == 3)
-					return new Vector3D((double)vals[0], (double)vals[1], (double)vals[2]);
+					if (rv is double d)
+						vals.Add(d);
+					else if(rv is BigInteger bi)
+						vals.Add((double)bi);
+				}
+
+				if (vals.Count == 2)
+					return new Vector2D(vals[0], vals[1]);
+				if (vals.Count == 3)
+					return new Vector3D(vals[0], vals[1], vals[2]);
 			}
 			else if (data is ExpandoObject)
 			{
 				IDictionary<string, object> eo = data as IDictionary<string, object>;
 				double x = double.NaN;
 				double y = double.NaN;
-				double z = double.NaN;
+				double z = 0.0;
 
 				if (eo.ContainsKey("x"))
 				{
@@ -524,24 +541,11 @@ namespace CodeImp.DoomBuilder.UDBScript
 					}
 				}
 
-				if (allow3d)
-				{
-					if (!double.IsNaN(x) && !double.IsNaN(y) && double.IsNaN(z))
-						return new Vector2D(x, y);
-					else if (!double.IsNaN(x) && !double.IsNaN(y) && !double.IsNaN(z))
-						return new Vector3D(x, y, z);
-				}
-				else
-				{
-					if (x != double.NaN && y != double.NaN)
-						return new Vector2D(x, y);
-				}
+				if (!double.IsNaN(x) && !double.IsNaN(y) && !double.IsNaN(z))
+					return new Vector3D(x, y, z);
 			}
 
-			if (allow3d)
-				throw new CantConvertToVectorException("Data must be a Vector2D, Vector3D, or an array of numbers.");
-			else
-				throw new CantConvertToVectorException("Data must be a Vector2D, or an array of numbers.");
+			throw new CantConvertToVectorException("Data must be a Vector2D, Vector3D, an array of numbers, or an object with (x, y, z) members.");
 		}
 
 		internal object GetConvertedUniValue(UniValue uv)
